pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mintdeck-localize"
version = "0.1.0"
description = "iOS localization pipeline for MintDeck"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.2",
    "deepl>=1.15.0",
    "openai>=1.3.0",
    "click>=8.1.7",
    "rich>=13.7.0",
]

[project.scripts]
localize = "src.cli:cli"
localize-web = "src.web.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]
web = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
    "aiofiles>=23.2.1",
]

[tool.setuptools.packages.find]
where = ["."]
